Extract Prodia models to a separate file
